trooper/members/managers.py
from django.contrib.auth.models import UserManager


class MemberManager(UserManager):
    def get_by_natural_key(self, username):
        return self.get(**{f"{self.model.USERNAME_FIELD}__iexact": username})
