requirements/local.txt
-r base.txt

black
flake8
isort
pre-commit
django-extensions
django-debug-toolbar
